Rename and refactor firebaseUtils class to firebaseService
